refactor: errror => triangle-alert로 변경
